StackOverflow
Stack Overflow is an application that displays questions and answers (using Entity Framework). A user can log in to ask, answer, and like a question. Any user, whether logged in or not, can view the questions and answers.
To get the project up and running: In your command line, navigate to the StackOverflow.Data folder. Once in the correct folder, type dotnet ef database update in the command line. This will create a database in SQL Server and then the project will be able to run.
Thank you!